ref(pack): make pack a chart subcommand
